move system_tray module mostly over
