+++ rust/gui-client/src-common/Cargo.toml
@@ -16,6 +16,7 @@ futures = { version = "0.3", default-features = false }
 hex = "0.4.3"
 minidumper = "0.8.3"
 output_vt100 = "0.1"
+png = "0.17.13" # `png` is mostly free since we already need it for Tauri
 rand = "0.8.5"
 reqwest = { version = "0.12.5", default-features = false, features = ["stream", "rustls-tls"] }
 secrecy = { workspace = true }
