+++ rust/gui-client/src-common/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod auth;
+pub mod compositor;
 pub mod crash_handling;
 pub mod deep_link;
 pub mod ipc;
 pub mod logging;
 pub mod settings;
+pub mod system_tray;
 pub mod updates;
